moved email templates. modified dockerfile to build templates first to be able to be embedded
