+++ Dockerfile
@@ -1,3 +1,12 @@
+# Build MJML email templates
+FROM node:23 AS mjml_builder
+WORKDIR /app
+COPY . .
+
+RUN npm i -g mjml
+RUN mjml internal/email/emails/*.mjml -o internal/email/emails/
+
+# Build API binary
 FROM golang:1.24-alpine AS build
 RUN apk add --no-cache alpine-sdk
 
@@ -6,7 +15,7 @@ WORKDIR /app
 COPY go.mod go.sum ./
 RUN go mod download
 
-COPY . .
+COPY --from=mjml_builder /app/. .
 
 RUN CGO_ENABLED=1 GOOS=linux go build -o main cmd/api/main.go
 
@@ -24,12 +33,6 @@ COPY frontend/package*.json ./
 RUN npm install
 COPY frontend/. .
 RUN npm run build
-
-# FROM node:23-slim AS frontend
-# RUN npm install -g serve
-# COPY --from=frontend_builder /frontend/dist /app/dist
-# EXPOSE 5173
-# CMD ["serve", "-s", "/app/dist", "-l", "5173"]
 
 FROM nginx:alpine AS frontend
 COPY --from=frontend_builder /frontend/dist /usr/share/nginx/html
